+++ CHANGES.rst
@@ -5,6 +5,7 @@ Changelog
 -------------------
 
 - dropped numpy<2.0.0 restriction
+- added `grayscale-to-binary` filter
 
 
 0.0.11 (2025-04-03)